pyproject.toml
[project]
name = "codeql-mcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastmcp>=2.12.4",
    "httpx>=0.28.1",
    "mcp[cli]",
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = ["-v"]

[tool.coverage.run]
source = ["server", "codeqlclient"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError"
]

[dependency-groups]
test = [
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]
